+++ apps/client/src/pages/my/hooks/use-performance-detail.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/confeti-detail/performance-queries';
+import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/performance/performance-queries';
 
 export const useUserPerformanceMore = (concertId: number) => {
   const { data } = useSuspenseQuery(
