+++ apps/client/src/pages/performance/hooks/use-concert-detail.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/confeti-detail/performance-queries';
+import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/performance/performance-queries';
 import { Concert, ConcertArtist } from '@shared/types/concert-response';
 
 interface UseConcertDetailReturn {
